🎨 헤더 분리 완료 #70
