+++ src/pages/viewPost/ViewPost.jsx
@@ -15,6 +15,7 @@ import { Helmet } from "react-helmet-async";
 import { viewPost } from "../../api/viewpostApi";
 import { myInfo } from "../../api/myInfoApi";
 import { uploadComment, getCommentList } from "../../api/commentApi";
+import HeaderText from "../../components/header/HeaderText";
 
 export default function ViewPost() {
   const { post_id } = useParams();
@@ -95,6 +96,7 @@ export default function ViewPost() {
       <Helmet>
         <title>Campic | 게시글 상세</title>
       </Helmet>
+      <HeaderText text={""} />
       <WrapViewPost>
         {data && <PostItem data={data} commentCount={comments.length} />}
         <CommentSection>
